// sourcecode/BTrak.Packages/snova-project-management/PM-module-components/pm-components/src/lib/snovasys-projects/models/templates-model.ts
export class TemplateModel {
    templateId : string;
    templateName : string;
    templateResponsibleUserId: string;
    templateResponsiblePersonId: string;
    templateResponsiblePersonName: string;
    templateResponsibleProfileImage: string;
    boardTypeId : string;
    workflowId: string;
    goalId: string;
    projectId: string;
    onBoardProcessDate: Date;
    createdDateTime: any;
    createdByUserId: string;
    timeStamp: any;
    isArchived: boolean;
}